util/bitmask: Move bitmask functionality its own file
