+++ source/common.hpp
@@ -48,6 +48,7 @@
 #include "config.hpp"
 #include "strings.hpp"
 #include "version.hpp"
+#include "util/util-bitmask.hpp"
 #include "util/util-library.hpp"
 #include "util/util-math.hpp"
 #include "util/util-profiler.hpp"
